Refactor extract and extract_all to work as extractors directly
These functions now take an optional parameter that, if present,
will change the return value to be the extracted data instead of a
callable extractor. This will require less setup for the times when the
extractor is called only once.

The cook has been changed to ignore if the input is already soup.